v28.x_29.x

+++ 1-install.sh
@@ -1,30 +1,55 @@
 #!/bin/bash
 
-extName=shamsi-calendar@gnome.scr.ir
+uuid=shamsi-calendar@gnome.scr.ir
 
 extFontsDirName=shamsiCalendarFonts
 
-cd $extName
+gnomeVersion=$( gnome-shell --version | sed -Ee 's/GNOME Shell (([0-9]{1,2}).([0-9]{1,2})*).*/\1/i;q' )
 
-## After Change & Develop:
-glib-compile-schemas ./schemas/
-zip -qrD9 ../extension.zip ./
+dirName=""
 
-## Install Extension:
-extDir=$HOME/.local/share/gnome-shell/extensions/$extName/
-rm -rf $extDir
-mkdir -p $extDir && cp -r ./* $extDir
+if [ "`echo "${gnomeVersion} >= 45" | bc`" -eq 1 ]
+then
+	dirName="gnome_45"
+elif [ "`echo "${gnomeVersion} >= 36" | bc`" -eq 1 -a "`echo "${gnomeVersion} < 45" | bc`" -eq 1 ]
+then
+	dirName="gnome_3.38-44"
+elif [ "`echo "${gnomeVersion} >= 24" | bc`" -eq 1 -a "`echo "${gnomeVersion} < 36" | bc`" -eq 1 ]
+then
+	dirName="gnome_3.24-3.34"
+elif [ "`echo "${gnomeVersion} >= 20" | bc`" -eq 1 -a "`echo "${gnomeVersion} < 24" | bc`" -eq 1 ]
+then
+	dirName="gnome_3.20-3.22"
+else
+	echo "Gnome $gnomeVersion not support! Please Update extension"
+fi
 
-## Install Fonts:
-extFontDir=$HOME/.local/share/fonts/$extFontsDirName/
-rm -rf $extFontDir
-mkdir -p $extFontDir && cp -r ./fonts/* $extFontDir
-
-echo ""
-echo "Please restart gnome-shell:"
-echo "Xorg => (Alt+F2 -> r -> Enter)"
-echo "Wayland [or Xorg] => (logOut + logIn)"
-echo "Then run ./2-install.sh"
-echo ""
-
-
+if [ $dirName != "" ]
+then
+	# echo "Gnome $gnomeVersion : ./extension/$dirName/"
+    
+    cd ./extension/$dirName/$uuid
+    
+    ## After Change & Develop:
+    glib-compile-schemas ./schemas/
+    zip -qrD9 ../$dirName.zip ./
+    
+    ## Install Extension:
+    extDir=$HOME/.local/share/gnome-shell/extensions/$uuid/
+    rm -rf $extDir
+    mkdir -p $extDir && cp -r ./* $extDir
+    
+    ## Install Fonts:
+    extFontDir=$HOME/.local/share/fonts/$extFontsDirName/
+    rm -rf $extFontDir
+    mkdir -p $extFontDir && cp -r ./fonts/* $extFontDir
+    
+    echo ""
+    echo "Please restart gnome-shell:"
+    echo " Wayland [or Xorg] => (logOut + logIn)"
+    echo " Xorg => (Alt+F2 -> r -> Enter)"
+    echo ""
+    echo "Then run ./2-install.sh"
+    echo ""
+        
+fi
